+++ sample-app/src/main/java/com/ryccoatika/simplesocket/ui/server/Server.kt
@@ -3,26 +3,41 @@ package com.ryccoatika.simplesocket.ui.server
 import android.util.Log
 import androidx.activity.compose.BackHandler
 import androidx.compose.foundation.layout.Column
+import androidx.compose.foundation.layout.Row
 import androidx.compose.foundation.layout.Spacer
 import androidx.compose.foundation.layout.height
 import androidx.compose.foundation.layout.padding
 import androidx.compose.foundation.rememberScrollState
+import androidx.compose.foundation.text.KeyboardActions
+import androidx.compose.foundation.text.KeyboardOptions
 import androidx.compose.foundation.verticalScroll
 import androidx.compose.material.icons.Icons
 import androidx.compose.material.icons.filled.ArrowBack
+import androidx.compose.material3.Divider
+import androidx.compose.material3.DropdownMenuItem
 import androidx.compose.material3.ExperimentalMaterial3Api
+import androidx.compose.material3.ExposedDropdownMenuBox
+import androidx.compose.material3.ExposedDropdownMenuDefaults
 import androidx.compose.material3.Icon
 import androidx.compose.material3.IconButton
+import androidx.compose.material3.OutlinedTextField
 import androidx.compose.material3.Scaffold
 import androidx.compose.material3.Text
+import androidx.compose.material3.TextButton
+import androidx.compose.material3.TextField
 import androidx.compose.material3.TopAppBar
 import androidx.compose.runtime.Composable
 import androidx.compose.runtime.LaunchedEffect
+import androidx.compose.runtime.getValue
 import androidx.compose.runtime.mutableStateListOf
 import androidx.compose.runtime.mutableStateMapOf
+import androidx.compose.runtime.mutableStateOf
 import androidx.compose.runtime.remember
+import androidx.compose.runtime.setValue
+import androidx.compose.ui.Alignment
 import androidx.compose.ui.Modifier
 import androidx.compose.ui.platform.LocalContext
+import androidx.compose.ui.text.input.ImeAction
 import androidx.compose.ui.tooling.preview.Preview
 import androidx.compose.ui.unit.dp
 import com.ryccoatika.simplesocket.ui.theme.AppTheme
@@ -39,6 +54,8 @@ fun Server(
     val context = LocalContext.current
     val ipAddress = remember { NetworkUtils.getLocalIpv4Address(context) }
     val connectedClients = remember { mutableStateListOf<Client>() }
+    var selectedClient by remember { mutableStateOf<Client?>(null) }
+    var sendMessage by remember { mutableStateOf("") }
     val incomingMessages = remember { mutableStateMapOf<Client, String>() }
     val socketServer = remember {
         SocketServer(1111)
@@ -104,10 +121,79 @@ fun Server(
             Spacer(Modifier.height(10.dp))
             Text(text = "Connected Clients:")
             connectedClients.forEach { client ->
-                Text(text = "address: ${client.hostAddress}")
-                Text(text = "port: ${client.port}")
-                Text(text = "localPort: ${client.localPort}")
+                Column {
+                    Spacer(Modifier.height(5.dp))
+                    Row(
+                        verticalAlignment = Alignment.CenterVertically,
+                    ) {
+                        Column {
+                            Text(text = "address: ${client.hostAddress}")
+                            Text(text = "port: ${client.port}")
+                            Text(text = "localPort: ${client.localPort}")
+                        }
+                        TextButton(
+                            onClick = {
+                                socketServer.disconnectClient(client)
+                            }
+                        ) {
+                            Text("Disconnect")
+                        }
+                    }
+                    Spacer(Modifier.height(5.dp))
+                    Divider()
+                }
             }
+            Spacer(Modifier.height(10.dp))
+            var showDropdown by remember { mutableStateOf(false) }
+            ExposedDropdownMenuBox(
+                expanded = showDropdown,
+                onExpandedChange = {
+                    showDropdown = !showDropdown
+                }
+            ) {
+                TextField(
+                    value = selectedClient?.hostAddress ?: "",
+                    onValueChange = {},
+                    readOnly = true,
+                    trailingIcon = { ExposedDropdownMenuDefaults.TrailingIcon(expanded = showDropdown) },
+                    modifier = Modifier.menuAnchor()
+                )
+                ExposedDropdownMenu(
+                    expanded = showDropdown,
+                    onDismissRequest = { showDropdown = !showDropdown }
+                ) {
+                    connectedClients.forEach { client ->
+                        DropdownMenuItem(
+                            text = {
+                                Text(client.hostAddress)
+                            },
+                            onClick = {
+                                selectedClient = client
+                                showDropdown = !showDropdown
+                            }
+                        )
+                    }
+                }
+            }
+            OutlinedTextField(
+                value = sendMessage,
+                label = {
+                    Text(text = "enter message")
+                },
+                onValueChange = {
+                    sendMessage = it
+                },
+                keyboardOptions = KeyboardOptions(
+                    imeAction = ImeAction.Send,
+                ),
+                keyboardActions = KeyboardActions(
+                    onSend = {
+                        selectedClient?.let {
+                            socketServer.sendMessage(selectedClient, sendMessage)
+                        }
+                    }
+                )
+            )
             Spacer(Modifier.height(10.dp))
             Text(text = "Messages:")
             incomingMessages.keys.forEach { client ->